Renaming channels.notifications Get/Post endpoints to subscriptions.getOne and rooms.saveNotifications
