+++ tests/end-to-end/api/09-rooms.js
@@ -1,7 +1,7 @@
 /* eslint-env mocha */
 /* globals expect */
 
-import {getCredentials, api, request, credentials } from '../../data/api-data.js';
+import { getCredentials, api, request, credentials } from '../../data/api-data.js';
 
 describe('[Rooms]', function() {
 	this.retries(0);
@@ -33,5 +33,42 @@ describe('[Rooms]', function() {
 				expect(res.body).to.have.property('remove').that.have.lengthOf(0);
 			})
 			.end(done);
+	});
+
+	describe('/rooms.saveNotification:', () => {
+		let testChannel;
+		it('create an channel', (done) => {
+			request.post(api('channels.create'))
+				.set(credentials)
+				.send({
+					name: `channel.test.${ Date.now() }`
+				})
+				.end((err, res) => {
+					testChannel = res.body.channel;
+					done();
+				});
+		});
+		it('/rooms.saveNotification:', (done) => {
+			request.post(api('rooms.saveNotification'))
+				.set(credentials)
+				.send({
+					roomId: testChannel._id,
+					notifications: {
+						disableNotifications: '0',
+						emailNotifications: 'nothing',
+						audioNotificationValue: 'beep',
+						desktopNotifications: 'nothing',
+						desktopNotificationDuration: '2',
+						audioNotifications: 'all',
+						mobilePushNotifications: 'mentions'
+					}
+				})
+				.expect('Content-Type', 'application/json')
+				.expect(200)
+				.expect((res) => {
+					expect(res.body).to.have.property('success', true);
+				})
+				.end(done);
+		});
 	});
 });
